Add collision detection and boundary shading
Introduces collision detection for player movement to prevent walking through walls. Adds boundary shading logic to improve wall rendering, and displays player position and map on the screen. Will add comments in the next commit to better under stand what going on. Also want to see if i ca implament mouse movement to look left and right and change A and S to move left and right insead of viewing