Memory Leak
Fixed some memory leak probelms. Wanted to also comment on the past upload that I want to add agoal to the game where player must look for a tile or wall or something in order to win the game. They will be timed to find this thing.

Was also thinking of flashing wall yellow when timer is nearing its end and red when its about to go off